Refactored file structure, created api folder, enhanced linting for js files
