+++ src/components/NavItems/tools/Keyframes/Hooks/useKeyframeWrapper.jsx
@@ -3,10 +3,16 @@ import { useState } from "react";
 import { useDispatch } from "react-redux";
 
 import {
+  useCreateKeyframeJob,
+  useFetchKeyframeFeatures,
+  useFetchKeyframes,
+  usePollKeyframeJobStatus,
+} from "@/components/NavItems/tools/Keyframes/api";
+import { KeyframeInputType } from "@/components/NavItems/tools/Keyframes/api/createKeyframeJob";
+import {
   setKeyframesFeatures,
   setKeyframesResult,
 } from "@/redux/reducers/tools/keyframesReducer";
-import useAuthenticatedRequest from "@Shared/Authentication/useAuthenticatedRequest";
 import { isValidUrl } from "@Shared/Utils/URLUtils";
 
 export const useProcessKeyframes = () => {
@@ -15,7 +21,10 @@ export const useProcessKeyframes = () => {
 
   const dispatch = useDispatch();
 
-  const authenticatedRequest = useAuthenticatedRequest();
+  const createKeyframeJob = useCreateKeyframeJob();
+  const pollKeyframeJobStatus = usePollKeyframeJobStatus(setStatus);
+  const fetchKeyframeFeatures = useFetchKeyframeFeatures();
+  const fetchKeyframes = useFetchKeyframes();
 
   // Step 1: Send URL
   const sendUrlMutation = useMutation({
@@ -27,86 +36,14 @@ export const useProcessKeyframes = () => {
 
       setStatus("Sending URL...");
 
-      const d = JSON.stringify({
-        video: url,
-      });
-
-      const config = {
-        method: "post",
-        maxBodyLength: Infinity,
-        url: process.env.REACT_APP_KEYFRAME_API_2,
-        headers: {
-          "Content-Type": "application/json",
-        },
-        data: d,
-        timeout: 10000, //We are setting a 10sec timeout
-      };
-      const response = await authenticatedRequest(config);
-
-      //job id validation
-      if (
-        !response.data ||
-        !response.data.session ||
-        typeof response.data.session !== "string" ||
-        response.data.session === ""
-      ) {
-        throw new Error("Invalid KSE session");
-      }
-
-      return response.data.session;
+      return await createKeyframeJob(KeyframeInputType.URL, url);
     },
   });
 
   // Step 2: Polling for status
   const checkStatusMutation = useMutation({
     mutationFn: async (jobId) => {
-      let currentStatus;
-
-      // Sometimes the API can return an empty status. We retry polling the status up to 3 times.
-      let emptyStatusTriesNumber = 0;
-      const EMPTY_STATUS_MAX_TRIES = 3;
-
-      do {
-        if (emptyStatusTriesNumber > EMPTY_STATUS_MAX_TRIES) {
-          throw new Error("Empty status. Try again.");
-        }
-
-        const config = {
-          method: "get",
-          url: `${process.env.REACT_APP_KEYFRAME_API_2}/status/${jobId}`,
-        };
-
-        const response = await authenticatedRequest(config);
-
-        if (!response.data.status) {
-          throw new Error("KSE Job id not found");
-        }
-
-        currentStatus = response.data.status;
-
-        const statusMessage =
-          currentStatus.indexOf(":") !== -1
-            ? currentStatus.substring(0, currentStatus.indexOf(":"))
-            : currentStatus;
-        const statusPercentage = currentStatus.split(":").pop();
-
-        if (statusMessage.includes("failed")) {
-          throw new Error(statusMessage);
-        }
-
-        if (!statusMessage) {
-          emptyStatusTriesNumber++;
-          setStatus(`Processing... Trying to retrieve the status`);
-        } else {
-          setStatus(`Processing... ${statusMessage} ${statusPercentage}%`);
-        }
-
-        if (currentStatus !== "completed:::100") {
-          await new Promise((resolve) => setTimeout(resolve, 1000)); // Wait 3s before next check
-        }
-      } while (currentStatus !== "completed:::100");
-
-      return jobId;
+      return await pollKeyframeJobStatus(jobId);
     },
   });
 
@@ -114,153 +51,20 @@ export const useProcessKeyframes = () => {
   const fetchFeatureDataMutation = useMutation({
     mutationFn: async (jobId) => {
       setStatus("Retrieving features...");
-      const config = {
-        method: "get",
-        url: `${process.env.REACT_APP_KEYFRAME_API_2}/keyframes_enhance/${jobId}`,
-      };
-
-      const response = await authenticatedRequest(config);
-      let faces = /** @type {ImagesFeature[]} */ [];
-      let texts = /** @type {ImagesFeature[]} */ [];
-
-      for (const imageFeature of response.data.faces_groups) {
-        const items = /** @type {ImageItem[]} */ [];
-
-        for (const it of imageFeature.items) {
-          items.push({
-            frame: it.frame,
-            frameTime: it.frame_time,
-            imageUrl: it.item_url,
-          });
-        }
-
-        const representative = /** @type {ImageRepresentative} */ {
-          index: imageFeature.representative.item_index,
-          imageUrl: imageFeature.representative.enhanced_url,
-        };
-
-        faces.push({
-          items: items,
-          representative: representative,
-        });
-      }
-
-      for (const imageFeature of response.data.text_groups) {
-        const items = /** @type {ImageItem[]} */ [];
-
-        for (const it of imageFeature.items) {
-          items.push({
-            frame: it.frame,
-            frameTime: it.frame_time,
-            imageUrl: it.item_url,
-          });
-        }
-
-        const representative = /** @type {ImageRepresentative} */ {
-          index: imageFeature.representative.item_index,
-          imageUrl: imageFeature.representative.enhanced_url,
-        };
-
-        texts.push({
-          items: items,
-          representative: representative,
-        });
-      }
-
-      setFeatureStatus("Completed");
-
-      return /** @type {KeyframesFeatures} */ {
-        faces: faces,
-        texts: texts,
-      };
+      return await fetchKeyframeFeatures(jobId);
     },
     onSuccess: (data) => {
+      setFeatureStatus("Completed");
       dispatch(setKeyframesFeatures(data));
     },
   });
 
   const fetchDataMutation = useMutation({
     mutationFn: async (jobId) => {
-      const config = {
-        method: "get",
-        url: `${process.env.REACT_APP_KEYFRAME_API_2}/result/${jobId}`,
-      };
-
-      const response = await authenticatedRequest(config);
-      setStatus("Completed");
-
-      let shots = /** @type {Shot[]} */ [];
-
-      for (const shot of response.data.shots) {
-        shots.push({
-          shotNumber: response.data.shots.indexOf(shot),
-          beginFrame: shot.beginframe,
-          beginTime: shot.begintime,
-          endFrame: shot.endframe,
-          endTime: shot.endtime,
-        });
-      }
-
-      let subshots = /** @type {Subshot[]} */ [];
-
-      for (const subshot of response.data.subshots) {
-        subshots.push({
-          subshotNumber: response.data.subshots.indexOf(subshot),
-          beginFrame: subshot.beginframe,
-          beginTime: subshot.begintime,
-          endFrame: subshot.endframe,
-          endTime: subshot.endtime,
-          shot: subshot.shot,
-        });
-      }
-
-      let keyframes = /** @type {Keyframe[]} */ [];
-      for (const kf of response.data.keyframes) {
-        keyframes.push({
-          frame: kf.frame,
-          keyframeTime: kf.keyframe_time,
-          keyframeUrl: kf.keyframe_url,
-          shot: kf.shot,
-          subshot: kf.subshot,
-        });
-      }
-
-      let keyframesXtra = /** @type {Keyframe[]} */ [];
-      for (const kf of response.data.keyframes_xtra) {
-        keyframesXtra.push({
-          frame: kf.frame,
-          keyframeTime: kf.keyframe_time,
-          keyframeUrl: kf.keyframe_url,
-          shot: kf.shot,
-          subshot: kf.subshot,
-        });
-      }
-
-      return /** @type {KeyframesData} */ {
-        session:
-          typeof response.data.session === "string"
-            ? response.data.session
-            : "",
-        url: typeof response.data.url === "string" ? response.data.url : "",
-        duration:
-          typeof response.data.duration === "string"
-            ? response.data.duration
-            : "",
-        framerate:
-          typeof response.data.framerate === "number"
-            ? response.data.framerate
-            : 0,
-        keyframes: keyframes,
-        keyframesXtra: keyframesXtra,
-        shots: shots,
-        subshots: subshots,
-        zipFileUrl:
-          typeof response.data.keyframes_zip === "string"
-            ? response.data.keyframes_zip
-            : "",
-      };
+      return await fetchKeyframes(jobId);
     },
     onSuccess: (data) => {
+      setStatus("Completed");
       dispatch(setKeyframesResult(data));
     },
   });
